refactor(ContextMenu): absorb ContextMenuView
