+++ packages/code-editor/src/ContextMenu/ContextMenu.styles.ts
@@ -3,3 +3,13 @@ import { css } from '@leafygreen-ui/emotion';
 export const containerStyles = css`
   display: contents;
 `;
+
+export const getMenuContainerStyles = (position: {
+  x: number;
+  y: number;
+}) => css`
+  position: absolute;
+  z-index: 50;
+  top: ${position.y}px;
+  left: ${position.x}px;
+`;
